--- .github/CONTRIBUTING.md
@@ -1,11 +1,0 @@
-## Contribuindo
-
-### Pull request
-
-Os pull requests devem seguir o template pré-gerado, devendo ter as características abaixo:
-  - Arquivo de dependências (`package.json`, `build.gradle`, etc.)
-  - Os arquivos de tal projeto deverão estar no caminho correto: `nome da linguagem/nome do framework (se existir)/nome do projeto`. Ex.: `javascript/discord.js/bot-simples`
-
-### Issues
-
-Os issues devem seguir o template pré-gerado. Caso resolva ignorar o template, você deve explicar o seu problema de maneira completa e entendível.
