--- .github/ISSUE_TEMPLATE.md
@@ -1,11 +1,0 @@
-### Qual o problema?
-
-  - Digite aqui o problema.
-
-### O problema é em determinado projeto? Se sim, qual?
-  
-  - Caso o seu problema seja com algum projeto, digite-o aqui.
-
-### Como reproduzir esse problema?
-
-  - Digite aqui os passos para reproduzir esse problema.
