--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,9 +1,0 @@
-### O que esse pull request adiciona?
-
-  - Digite aqui o que o seu pull request adiciona ao repositório.
-
-### Tipo de pull request
-  
-  - Coloque um `x` dentro dos colchetes para marcar o tipo do seu pull request.
-  - [ ] Melhoria do repositório
-  - [ ] Novo projeto
