--- README.md
@@ -1,8 +1,0 @@
-# ![BR PROGRAMMERS](https://i.imgur.com/yeUW4BV.png) BR PROGRAMMERS
-[![Discord](https://img.shields.io/discord/278636362163552257.svg)](https://discord.gg/8kwH355)
-
-Repositório do servidor de programação no Discord, BR PROGRAMMERS. Tem como foco reunir projetos dos membros de lá, para ajudar os novatos e/ou já experientes. 
-
-## Como posso contribuir?
-
-Você pode ler as regras de contribuição [aqui](https://github.com/BestVelcro/BR-PROGRAMMERS/blob/master/.github/CONTRIBUTING.md)!
