+++ discord.js/bot-avançado/comandos/avatar.js
