+++ discord.js/bot-avançado/eventos/ready.js
@@ -1,0 +1,3 @@
+exports.run = (client) => {
+  console.log('Pronto! Autenticado como ' + client.user.username);
+}
