+++ discord.js/primeiro-bot/package.json
