+++ eris/bot simples/bot.js
