+++ eris/bot simples/package.json
