--- javascript/discord.js/bot-avançado/eventos/ready.js
@@ -1,3 +1,0 @@
-exports.run = (client) => {
-  console.log('Pronto! Autenticado como ' + client.user.username);
-}
