--- python/discord.py/basic-bot/requirements.txt
@@ -1,1 +1,0 @@
-discord.py
